Added regions methods

+++ if_else_2024/accounts/services.py
@@ -6,12 +6,17 @@ from if_else_2024.auth.utils import pass_context
 from if_else_2024.core.exceptions import (
     EntityAlreadyExistsException,
     EntityNotFoundException,
+    IntegrityBreachException,
 )
+from if_else_2024.regions.repositories import RegionRepository
 
 
 class AccountService:
-    def __init__(self, account_repository: AccountRepository):
+    def __init__(
+        self, account_repository: AccountRepository, region_repository: RegionRepository
+    ):
         self._repository = account_repository
+        self._region_repository = region_repository
 
     async def get_by_id(self, session: AsyncSession, id: int):
         account = await self._repository.get_by_id(session, id)
@@ -57,5 +62,8 @@ class AccountService:
         account = await self._repository.get_by_id(session, id)
         if account is None:
             raise EntityNotFoundException("Account with given id was not found")
+
+        if await self._region_repository.exists_by_account_id(session, id):
+            raise IntegrityBreachException("Account is creator of some regions")
 
         await self._repository.delete(session, account)
